+++ app/src/main/java/com/aisleron/ui/shoppinglist/ShoppingListItemRecyclerViewAdapter.kt
@@ -50,7 +50,7 @@ class ShoppingListItemRecyclerViewAdapter(
     }
 
     override fun getItemId(position: Int): Long {
-        return values[position].id
+        return values[position].id.toLong()
     }
 
     override fun getItemViewType(position: Int): Int {
